backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
redis==4.5.1
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==2.3.6
celery==5.3.0
cryptography==41.0.3
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
